+++ src/spec/EntitySpec.ts
@@ -56,6 +56,21 @@ describe('An entity', () => {
     expect(entity.types).toEqual([]);
   });
 
+  it('can get a list of components', () => {
+    const entity = new ex.Entity();
+    const typeA = new FakeComponent('A');
+    const typeB = new FakeComponent('B');
+    const typeC = new FakeComponent('C');
+
+    expect(entity.getComponents()).toEqual([]);
+
+    entity.addComponent(typeA)
+      .addComponent(typeB)
+      .addComponent(typeC);
+
+    expect(entity.getComponents().sort((a, b) => a.type.localeCompare(b.type))).toEqual([typeA, typeB, typeC]);
+  });
+
   it('can have type from tag components', () => {
     const entity = new ex.Entity();
     const isOffscreen = new TagComponent('offscreen');
@@ -73,7 +88,7 @@ describe('An entity', () => {
   it('can be observed for added changes', (done) => {
     const entity = new ex.Entity();
     const typeA = new FakeComponent('A');
-    entity.changes.register({
+    entity.componentAdded$.register({
       notify: (change) => {
         expect(change.type).toBe('Component Added');
         expect(change.data.entity).toBe(entity);
@@ -89,7 +104,7 @@ describe('An entity', () => {
     const typeA = new FakeComponent('A');
 
     entity.addComponent(typeA);
-    entity.changes.register({
+    entity.componentRemoved$.register({
       notify: (change) => {
         expect(change.type).toBe('Component Removed');
         expect(change.data.entity).toBe(entity);
@@ -107,14 +122,44 @@ describe('An entity', () => {
     const typeB = new FakeComponent('B');
     entity.addComponent(typeA);
     entity.addComponent(typeB);
+    entity.addChild(
+      new ex.Entity([
+        new FakeComponent('Z')
+      ])
+    );
 
     const clone = entity.clone();
     expect(clone).not.toBe(entity);
     expect(clone.id).not.toBe(entity.id);
-    expect(clone.components.A).not.toBe(entity.components.A);
-    expect(clone.components.B).not.toBe(entity.components.B);
-    expect(clone.components.A.type).toBe(entity.components.A.type);
-    expect(clone.components.B.type).toBe(entity.components.B.type);
+    expect(clone.get('A')).not.toBe(entity.get('A'));
+    expect(clone.get('B')).not.toBe(entity.get('B'));
+    expect(clone.get('A').type).toBe(entity.get('A').type);
+    expect(clone.get('B').type).toBe(entity.get('B').type);
+    expect(clone.children.length).toBe(1);
+    expect(clone.children[0].types).toEqual(['Z']);
+  });
+
+  it('can be initialized with a template', () => {
+    const entity = new ex.Entity();
+    const template = new ex.Entity([
+      new FakeComponent('A'),
+      new FakeComponent('B')
+    ]).addChild(
+      new ex.Entity([
+        new FakeComponent('C'),
+        new FakeComponent('D')
+      ]).addChild(
+        new ex.Entity([
+          new FakeComponent('Z')
+        ])
+      )
+    );
+
+    expect(entity.getComponents()).toEqual([]);
+    entity.addTemplate(template);
+    expect(entity.types.sort((a, b) => a.localeCompare(b))).toEqual(['A', 'B']);
+    expect(entity.children[0].types.sort((a, b) => a.localeCompare(b))).toEqual(['C', 'D']);
+    expect(entity.children[0].children[0].types).toEqual(['Z']);
   });
 
   it('can be checked if it has a component', () => {
@@ -186,7 +231,7 @@ describe('An entity', () => {
   it('can be parented', () => {
     const parent = new ex.Entity();
     const child = new ex.Entity();
-    parent.add(child);
+    parent.addChild(child);
 
     expect(child.parent).toBe(parent);
     expect(parent.children).toEqual([child]);
@@ -197,7 +242,7 @@ describe('An entity', () => {
     const parent = new ex.Entity();
     const child = new ex.Entity();
     const grandchild = new ex.Entity();
-    parent.add(child.add(grandchild));
+    parent.addChild(child.addChild(grandchild));
 
     expect(grandchild.parent).toBe(child);
     expect(child.parent).toBe(parent);
@@ -211,7 +256,7 @@ describe('An entity', () => {
     const parent = new ex.Entity();
     const child = new ex.Entity();
     const grandchild = new ex.Entity();
-    parent.add(child.add(grandchild));
+    parent.addChild(child.addChild(grandchild));
 
     expect(child.parent).toBe(parent);
 
@@ -225,10 +270,10 @@ describe('An entity', () => {
     const parent = new ex.Entity();
     const child = new ex.Entity();
 
-    parent.add(child);
+    parent.addChild(child);
 
     expect(() => {
-      child.add(parent);
+      child.addChild(parent);
     }).toThrowError('Cycle detected, cannot add entity');
   });
 
@@ -237,9 +282,9 @@ describe('An entity', () => {
     const child = new ex.Entity();
     const otherparent = new ex.Entity();
 
-    parent.add(child);
+    parent.addChild(child);
     expect(() => {
-      otherparent.add(child);
+      otherparent.addChild(child);
     }).toThrowError('Entity already has a parent, cannot add without unparenting');
   });
 
@@ -271,7 +316,7 @@ describe('An entity', () => {
     const e = new ex.Entity();
     const child = new ex.Entity();
     const grandchild = new ex.Entity();
-    e.add(child.add(grandchild));
+    e.addChild(child.addChild(grandchild));
 
     const world = new ex.World(new ex.Scene());
 
@@ -279,11 +324,11 @@ describe('An entity', () => {
 
     expect(world.entityManager.entities.length).toBe(3);
 
-    grandchild.add(new ex.Entity());
+    grandchild.addChild(new ex.Entity());
 
     expect(world.entityManager.entities.length).toBe(4);
 
-    child.remove(grandchild);
+    child.removeChild(grandchild);
 
     expect(world.entityManager.entities.length).toBe(2);
   });